update migrations, models, factory and seed

+++ database/migrations/2022_10_09_220908_create_brands_table.php
@@ -6,26 +6,26 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
+    public function up(): void
     {
         Schema::create('brands', function (Blueprint $table) {
             $table->id();
+            $table->string('title');
+
+            $table->string('slug')
+                ->unique();
+
+            $table->string('thumbnail')
+                ->nullable();
+
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
+    public function down(): void
     {
-        Schema::dropIfExists('brands');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('brands');
+        }
     }
 };
