+++ database/migrations/2022_10_12_105322_create_roles_table.php
@@ -1,32 +1,42 @@
 <?php
 
+use App\Models\Role;
+use App\Models\User;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
+    public function up(): void
     {
         Schema::create('roles', function (Blueprint $table) {
             $table->id();
-            $table->string('name')->unique();
+
+            $table->string('name')
+                ->unique();
+
+            $table->timestamps();
+        });
+
+        Schema::create('role_user', function (Blueprint $table) {
+            $table->id();
+
+            $table->foreignIdFor(Role::class)
+                ->constrained();
+
+            $table->foreignIdFor(User::class)
+                ->constrained();
+
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
+    public function down(): void
     {
-        Schema::dropIfExists('roles');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('roles');
+            Schema::dropIfExists('role_user');
+        }
     }
 };
